@@ -1,15 +1,17 @@
 package com.github.yvasyliev.config.app;
 
+import com.github.yvasyliev.model.entity.Tweet;
 import com.github.yvasyliev.model.entity.user.Role;
 import com.github.yvasyliev.model.entity.user.User;
+import com.github.yvasyliev.repository.TweetRepository;
 import com.github.yvasyliev.repository.UserRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
-import javax.transaction.Transactional;
 import java.time.LocalDateTime;
 
 @Component
@@ -20,10 +22,13 @@ public class ApplicationRunnerImpl implements ApplicationRunner {
     @Autowired
     private PasswordEncoder passwordEncoder;
 
+    @Autowired
+    private TweetRepository tweetRepository;
+
     @Transactional
     @Override
     public void run(ApplicationArguments args) {
-        userRepository.save(buildUser(
+        User johnDoe = userRepository.save(buildUser(
                 "admin",
                 "admin",
                 "john.doe@gmail.com",
@@ -31,6 +36,8 @@ public class ApplicationRunnerImpl implements ApplicationRunner {
                 "Doe",
                 Role.ADMIN
         ));
+
+        tweetRepository.save(buildTweet(johnDoe, "Hello World!"));
 
         userRepository.save(buildUser(
                 "user",
@@ -52,5 +59,12 @@ public class ApplicationRunnerImpl implements ApplicationRunner {
         user.setRole(role);
         user.setCreatedAt(LocalDateTime.now());
         return user;
+    }
+
+    private Tweet buildTweet(User user, String text) {
+        Tweet tweet = new Tweet();
+        tweet.setUser(user);
+        tweet.setText(text);
+        return tweet;
     }
 }
