@@ -1,18 +1,21 @@
 package com.github.yvasyliev.config.security;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.ForwardAuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @Component
 public class CustomForwardAuthenticationSuccessHandler extends ForwardAuthenticationSuccessHandler {
+    private final String forwardUrl;
+
     public CustomForwardAuthenticationSuccessHandler(@Value("/") String forwardUrl) {
         super(forwardUrl);
+        this.forwardUrl = forwardUrl;
     }
 
     @Override
@@ -21,7 +24,7 @@ public class CustomForwardAuthenticationSuccessHandler extends ForwardAuthentica
 
         response.sendRedirect(
                 redirectTo == null || redirectTo.isEmpty()
-                        ? "/"
+                        ? forwardUrl
                         : redirectTo
         );
     }
