@@ -4,6 +4,7 @@ import com.github.yvasyliev.model.dto.RegistrationForm;
 import com.github.yvasyliev.model.entity.user.User;
 import com.github.yvasyliev.model.event.UserRegisteredEvent;
 import com.github.yvasyliev.service.UserService;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Controller;
@@ -11,8 +12,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-
-import javax.servlet.http.HttpServletRequest;
 
 @Controller
 @RequestMapping(path = "/registration")
